third_party/blink/web_tests/wpt_internal/attribution-reporting/source-priority.sub.https.html

<!doctype html>
<meta charset=utf-8>
<meta name=timeout content=long>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>
<script>
attribution_reporting_promise_test(async t => {
  const sources = [{
      source_event_id: generateSourceEventId(),
      destination: 'https://{{host}}',
      priority: '11',
    },
    {
      source_event_id: generateSourceEventId(),
      destination: 'https://{{host}}',
      priority: '13',
    },
    {
      source_event_id: generateSourceEventId(),
      destination: 'https://{{host}}',
      priority: '12',
    }]
  await Promise.all(sources.map(source => registerAttributionSrc(t, {source})));
  await Promise.all(sources.map(source => waitForSourceToBeRegistered(source.source_event_id)));

  await registerAttributionSrc(t, {trigger: {
    event_trigger_data: [{}],
  }});

  const payload = await pollEventLevelReports();
  assert_equals(payload.reports.length, 1);
  const report = JSON.parse(payload.reports[0].body);
  assert_equals(report.source_event_id, sources[1].source_event_id);
}, 'Trigger is attributed to highest priority source.');
</script>
